pyproject.toml
[tool.poetry]
name = "emea3-capstone-ml"
version = "0.1.0"
description = "predicting price of used cars"
authors = ["charles.doyne <charles.doyne@datatonic.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
seaborn = "^0.13.2"
ipython = "^8.30.0"
ipykernel = "^6.29.5"
google-cloud-bigquery = {extras = ["pandas"], version = "^3.27.0"}
pandas-gbq = "^0.24.0"
db-dtypes = "^1.3.1"
numpy = "^2.1.3"
scikit-learn = "^1.5.2"
python-dotenv = "^1.0.1"
tqdm = "^4.67.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
